pyproject.toml
[tool.poetry]
name = "awsservices"
version = "0.1.0"
description = ""
authors = ["Vimit <talk2vimit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.2"
django-restframework = "^0.0.1"
Markdown = "^3.5.2"
Jinja2 = "^3.1.3"
psycopg-binary = "^3.1.18"
psycopg2-binary = "^2.9.9"
django-filter = "^23.5"
black = "^24.2.0"
isort = "^5.13.2"
django-guardian = "^2.4.0"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
pre-commit = "^3.6.2"
pillow = "^10.2.0"
jsonschema = "^4.21.1"
oauthlib = "^3.2.2"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
